requirements.txt
# Core
numpy~=2.0.2
pandas~=2.3.3
scikit-learn~=1.6.1
matplotlib~=3.9.4
tqdm~=4.67.1

# PyTorch (CLIP + Whisper backend)
torch~=2.8.0
torchvision
torchaudio

# OpenAI CLIP
git+https://github.com/openai/CLIP.git

# OpenAI Whisper
openai-whisper~=20250625

# Video & Audio Processing
opencv-python~=4.12.0.88
ffmpeg-python
librosa
soundfile

# Utility
requests
scipy~=1.13.1
pillow~=11.3.0
xgboost~=2.1.4
shap~=0.49.1
umap-learn~=0.5.9.post2
hdbscan~=0.8.40